requirements.txt
dataiku-api-client>=14.0.0
python-dotenv>=1.0.0
